+++ student-service/src/main/java/com/project/studentservice/repository/StudentRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
     Optional<Student> findByEmail(String email);
     Optional<Student> findByCardId(String cardId);
-    List<Student> findByBatchId(int batchId);
+    List<Student> findByBatchId(Long batchId);
     List<Student> findByBatchSection(char section);
 }
